Bug 1025 - GlueGen: Add accessor for compound fields of type array, pointer and string (code generation)
Enhance compound access as delivered by Bug 1022,
to also generate accessors (getter and setter) for
array, pointer and string types.

Allow configuration of array length either via
their internal size (c-header) or config 'ReturnedArrayLength'.

'ReturnedArrayLength' allows specifying a java expression.

Canonical field names of compounds are _now_ specified as
follows for configuration entries:
  COMPOUND.FIELD
e.g.
  StructA.fieldB

Also allow configuration of pointer fields to be treated as
referenced arrays via 'ReturnedArrayLength'.

Further, allow specifying 'pointer fields' as String values
via 'ReturnsString' configuration.

++++

Implementation details:
  - handle above described accessor features
  - enhance JavaDoc for generated accessors
  - generate native JNI compound and string accessor on demand
  - encapsule accessor code generation in their own methods
  - enhance exception messages
  - enhance type verbosity in debug mode
  - verbose debug output via GlueGen.debug()

Tests:
  - Features covered by test1.[ch]
    and Test1p1JavaEmitter and Test1p2ProcAddressEmitter

  - Validated compilation and unit tests for modules:
    - joal
    - jogl (minor config changes req.)
    - jocl (minor config changes req.)

+++ src/java/com/jogamp/common/nio/StructAccessor.java
@@ -117,7 +117,7 @@ public class StructAccessor {
     public final int getIntAt(int byteOffset, int nativeSizeInBytes) {
         switch(nativeSizeInBytes) {
             case 2:
-                return (int) bb.getShort(byteOffset) & 0x0000FFFF ;
+                return bb.getShort(byteOffset) & 0x0000FFFF ;
             case 4:
                 return bb.getInt(byteOffset);
             case 8:
@@ -137,7 +137,7 @@ public class StructAccessor {
                 bb.putInt(byteOffset, v);
                 break;
             case 8:
-                bb.putLong(byteOffset, (long)v & 0x00000000FFFFFFFFL );
+                bb.putLong(byteOffset, v & 0x00000000FFFFFFFFL );
                 break;
             default:
                 throw new InternalError("invalid nativeSizeInBytes "+nativeSizeInBytes);
@@ -178,7 +178,7 @@ public class StructAccessor {
     public final long getLongAt(int byteOffset, int nativeSizeInBytes) {
         switch(nativeSizeInBytes) {
             case 4:
-                return (long) bb.getInt(byteOffset) & 0x00000000FFFFFFFFL;
+                return bb.getInt(byteOffset) & 0x00000000FFFFFFFFL;
             case 8:
                 return bb.getLong(byteOffset);
             default:
